Move getSaveIds from test method to DB method.
It simplifies export_game by removing erroneous reads.
